backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.1
langgraph>=0.0.10
qdrant-client>=1.7.0
httpx
pydantic>=2.0.0
pypdf>=3.0.0